+++ CHANGELOG.md
@@ -55,6 +55,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 #### Changed
 - Docker Compose uses pre-built images of microservices for easier installation (part of [#430 - Microservices are built as Docker images](https://github.com/shopsys/shopsys/pull/430))
+- [#447 - Unable to resolve domain error page](https://github.com/shopsys/shopsys/pull/447)
 
 ### [shopsys/shopsys]
 #### Added